DBZ-7996 Move ChangeEventSink class to debezium-core and use the interface for the MongoDB sink connector
relates to https://issues.redhat.com/browse/DBZ-7996
